+++ changelog.md
@@ -1,3 +1,6 @@
+### 4.1.1
+- Fixed the crash report being printed to the log twice on integrated server crashes
+- Fixed the crash report txt file missing information on dedicated server crashes
 ## 4.1.0
 [Crashy](https://crashy.net/) is now supported in 1.16.5.
 # 4.0.0